Show data to CallView
