+++ MtoM fake/ApplyCell.swift
@@ -13,15 +13,20 @@ class ApplyCell: UITableViewCell {
     var headerApplyView = UIView()
     var webButton = UIButton()
     var callButton = UIButton()
-    
-    var callView = CallView()
-    
+    var job : Job! {
+        didSet {
+            callView = CallView(job: job)
+        }
+    }
+    var callView : CallView!
+
     override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
         setLayoutOfHeaderCell()
         setLayoutOfApplyCell()
-        
+        contentView.backgroundColor = UIColor.headerColor()
     }
+    
     
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
