+++ MtoM fake/TextCallCell.swift
@@ -12,8 +12,8 @@ class TextCallCell: TextCell {
     
     override func layoutBodyCell() {
         super.layoutBodyCell()
-        viewText.mt_innerAlign(left: 0, top: 0, right: 0, bottom: 0)
-        
+        viewText.mt_InnerAlign(allSpace: 0)
+        contentView.backgroundColor = UIColor.whiteColor()
     }
     
 }
