+++ MtoM fake/TitleCallCell.swift
@@ -13,23 +13,23 @@ class TitleCallCell: TitleCell {
     var callView : CallView?
     
     override func layoutBodyCell() {
-        self.callView = CallView(frame: frame)
+        callView = CallView(frame: frame)
         layoutBodyCel()
         headerOfCell.removeFromSuperview()
+        contentView.backgroundColor = UIColor.redColor()
     }
 
     func layoutBodyCel(){
         self.backgroundColor = UIColor.redColor()
         self.addSubview(clouseButton)
-        clouseButton.mt_InnerAlign(PinPosition.MidRight, space: 8, size: CGSize(width: 20, height: 20))
-        clouseButton.layer.cornerRadius = 9
+        clouseButton.mt_InnerAlign(PinPosition.MidRight, space: 8, size: CGSize(width: 30, height: 30))
+        clouseButton.layer.cornerRadius = 14
         clouseButton.setImage(UIImage(named: "ic_close"), forState: UIControlState.Normal)
         clouseButton.backgroundColor = UIColor.whiteColor()
         clouseButton.setTitleColor(UIColor.redColor(), forState: UIControlState.Normal)
         clouseButton.addTarget(self, action: "hideCallViewCell", forControlEvents: UIControlEvents.TouchUpInside)
     }
     func hideCallViewCell() {
-        callView!.removeFromSuperview()
+        callView!.hideCallView()
     }
-    
 }
