+++ MtoM fake/TitleCell.swift
@@ -21,7 +21,7 @@ class TitleCell: UITableViewCell {
     }
     override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
-        
+        contentView.backgroundColor = UIColor.headerColor()
         self.addSubview(headerOfCell)
         self.addSubview(backgroundOfCell)
