+++ MtoM fake/WebTableView.swift
@@ -1,0 +1,13 @@
+//
+//  WebTableView.swift
+//  MtoM fake
+//
+//  Created by Student on 1/29/16.
+//  Copyright © 2016 Trương Thắng. All rights reserved.
+//
+
+import UIKit
+
+class WebTableView: TableView {
+
+}
